Fix migration to check for column existence

+++ db/migrate/20250329192134_remove_margin_from_app_configs.rb
@@ -1,5 +1,7 @@
-class RemoveMarginFromAppConfigs < ActiveRecord::Migration[7.1]
+class RemoveMarginFromAppConfigs < ActiveRecord::Migration[7.0]
   def change
-    remove_column :app_configs, :margin, :decimal
+    if column_exists?(:app_configs, :margin)
+      remove_column :app_configs, :margin, :decimal
+    end
   end
 end
